+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Bumped jora version to `1.0.0-beta.1`
+
 ## 1.0.0-beta.33 (07-05-2020)
 
 - Fixed `table`'s cell content rendering when col config is a string